Set up initial classes, service communication through queues
I created the required microservice classes and their methods according to
our project design. Implemented queues for the services to communicate through
and used objects to process data instead of json because that was easier to code
a working prototype with and json did not seem to have an extra use for now.
Might require rewriting to json communication in a later version.