Set up testcomms and util packages
I created a testcomms package to test code that will be used to convert
all service communication from queues to REST, to better suit the
microservices concept and make use of Callbacks. I also moved
the DataReply, DataRequest and RequestType classes to the util
package and created a Util.java file that contains utility methods.