Update PinService to work with pin database, untested
I updated the pin service to work with a pin database and added a createTable class that will
set up the database pin needs, however this database has not yet been created and thus it hasnt been
tested yet. But the pin service should now be able to check cardnumber/pin combinations and link
them to the customer that owns the card, which should make the pin service functional again.