Move TableCreator to util, update javadoc, update sql
I moved the TableCreator class to the util package and updated its javadoc. I also
updated the SQL drop statements for the tables so they first check if the tables exist.
This is to make sure an exception does not occur if we try to drop a table that does not exist.